requirements.txt
alembic == 1.18.4
fastapi == 0.128.8
pydantic[email] == 2.12.5
pwdlib[argon2] == 0.3.0
pyjwt == 2.11.0
sqlalchemy == 2.0.46
aiosqlite == 0.22.1
python-multipart == 0.0.22
uvicorn[standard] == 0.40.0
pydantic-settings == 2.12.0
